Create multilayer Tiff file with 107 variables from chelsa-climate.org

Gives the average values on the data recovered between 1981 and 2010. Monthly variables are average temperatures, min temperatures, max temperatures, precipitation, potential evapotranspiration (with Penman formula and with Thornthwaite formula), and total cloud cover. Others variables are climatic water deficit (with Penman and Thornthwaite), number of dry month (with Penman and Thornthwaite) and 19 bio variables (more information in chelsa documentation).

get_chelsa_current(
  extent,
  extent_latlon,
  EPSG,
  destination,
  resolution = 1000,
  rm_download = FALSE
)

Arguments

extent

character. First output of transform_shp_coutry_extent.

extent_latlon

character. Second output of transform_shp_country_extent

EPSG

int. to consider for this country/area.

destination

character. absolute path where to download files like here() output.

resolution

int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.

rm_download

boolean. If TRUE remove download files and folders. Keep only current_chelsa.tif in data_raw folder, default is FALSE.

Value

character. absolute path to current_chelsa.tif.

Details

resolution need to be carefully chosen because if Tiff file is too big, R can crash.

Unit of each climatic variable :

Name	Unit

Temperature average (tas)	°C x 10

Temperature min (tasmin)	°C x 10

Temperature max (tasmax)	°C x 10

Precipitation	kg.m^-2

Cloud area fraction (clt)	%

PET Penman	kg.m^-2

PET Thornthwaite	kg.m^-2

Climatic water deficit (Penman)	kg.m^-2

Climatic dater deficit (Thornthwaite)	kg.m^-2

Number of dry months (Penman)	month

Number of dry months (Thornthwaite)	month

bio1	°C x 10

bio2	°C x 10

bio3	°C x 10

bio4	°C x 10

bio5	°C x 10

bio6	°C x 10

bio7	°C x 10

bio8	°C x 10

bio9	°C x 10

bio10	°C x 10

bio11	°C x 10

bio12	kg.m^-2.year^-1

bio13	kg.m^-2.month^-1

bio14	kg.m^-2.month^-1

bio15	kg.m^-2

bio16	kg.m^-2.month^-1

bio17	kg.m^-2.month^-1

bio18	kg.m^-2.month^-1

bio19	kg.m^-2.month^-1