Create several multilayer Tiff files with 81 variables from chelsa-climate.org with future climat variables.

Gives predictions on future values for the choosen phase. Prediction are from multiple models (GFDL-ESM4, IPSL-CM6A-LR, MPI-ESM1-2-HR, MRI-ESM2-0, UKESM1-0-LL). Creates folder for each model and a folder with mean of this five models. Monthly variables are average temperatures, min temperatures, max temperatures, precipitation, potential evapotranspiration with Thornthwaite formula. Others variables are climatic water deficit with Thornthwaite, number of dry month with Thornthwaite and 19 bio variables (more information in chelsa documentation).

Usage

get_chelsa_future(
  extent,
  EPSG,
  destination,
  resolution = 1000,
  phase = "2071-2100",
  ssp = 585
)

Arguments

extent

character. First output of transform_shp_coutry_extend

EPSG

int. to consider for this country/area.

destination

character. absolute path where to download files like here() output.

resolution

int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.

phase

character. Must be in c("2041-2070", "2071-2100") match to years to download, default is "2071-2100".

ssp

character or int. Must be in c(126, 370, 585). It's the Shared Socio-economic Pathways you want to download.

Details

resolution need to be carefully choosen because if Tiff file is too big, R can crash.

Unit of each climatic variable :

Name	Unit

Temperature average (tas)	°C x 10

Temperature min (tasmin)	°C x 10

Temperature max (tasmax)	°C x 10

Precipitation	kg.m^-1

PET Thornthwaite	kg.m^-1

Climatic Water Deficit (Thornthwaite)	kg.m^-1

Number of dry month (Thornthwaite)	month

bio 1	°C x 10

bio 2	°C x 10

bio 3	°C x 10

bio 4	°C x 10

bio 5	°C x 10

bio 6	°C x 10

bio 7	°C x 10

bio 8	°C x 10

bio 9	°C x 10

bio 10	°C x 10

bio 11	°C x 10

bio 12	kg.m^-2.year^-1

bio 13	kg.m^-2.month^-1

bio 14	kg.m^-2.month^-1

bio 15	kg.m^-2

bio 16	kg.m^-2.month^-1

bio 17	kg.m^-2.month^-1

bio 18	kg.m^-2.month^-1

bio 19	kg.m^-2.month^-1