Create extent & extent in latitude and longitude

With a country name, a shapefile or a extent and an EPSG value. Succed to give extent in EPSG reproject and extent with default EPSG also knows as latitude and longitude coordinates.

transform_shp_country_extent(
  EPSG,
  country_name = NULL,
  shapefile_path = NULL,
  extent_project = NULL,
  extent_latlon = NULL
)

Arguments

EPSG

int. to consider for this country/area.

country_name

character. English country name, default is NULL.

shapefile_path

character. Path to a .shp file, default is NULL.

extent_project

int vector. length 4, with this order c(xmin, ymin, xmax, ymax), default is NULL.

extent_latlon

int vector. length 4, with this order c(lonmin, latmin, lonmax, latmax), default is NULL.

Value

character vector. with extent of the area in one character, other are latlon coord of area