Create multilayer Tiff file with 11 environmental variables

Variables are type of soil, elevation, slope, aspect, roughness, solar radiation, distance to sea, protected areas, distance to roads, distance to cities and town, distance to rivers & lake.

Usage

get_env_variables(
  extent_latlon,
  extent,
  EPSG,
  country_name,
  destination,
  resolution = 1000,
  rm_download = FALSE
)

Arguments

extent_latlon

int vector. in this order c(lon_min, lat_min, lon_max, lat_max).

extent

character. First output of transform_shp_country_extent function.

EPSG

int. to consider for this country/area.

country_name

character. country name (in english) which be use to collect protected areas. This country must be available in https://www.protectedplanet.net/en/thematic-areas/wdpa?tab=WDPA.

destination

character. absolute path where to download files like here() output.

resolution

int. in meters, recommended resolution are 250m, 500m, 1km, 2km or 5km, default is 1km. See more in details.

rm_download

boolean. If TRUE remove download files and folders. Keep only environ.tif in data_raw folder, default is FALSE.

Value

character. absolute path to environ.tif file.

Details

resolution need to be carefully choosen because if Tiff file is too big, R can crash.

environ.tif.aux.xml is an extention of environ.tif, it allows to classify soilgrid variable with QGIS with RasterAttributeTable extension. Nevertheless it's cause problems to open it with stars package but not with terra. If you have any problems to open environ.tif, you can remove environ.tif.aux.xml. This solve all accessibility problems with stars and terra packages.

Unit of each environ variable :

Name	Unit

Elevation	m

Aspect	degrees

Roughness	m

Slope	degrees

Solar radiation	Wh.m^-2.day

Soilgrids	category

Distance sea	m

Distance road	m

Distance place	m

Distance watering place	m

Protected Area (WDPA)	category